+++ example/lib/util/locale/localization.dart
@@ -10,32 +10,26 @@ import 'package:locale_gen_example/util/locale/localization_keys.dart';
 class Localization {
   Map<String, dynamic> _localisedValues = Map();
 
-  static Localization of(BuildContext context) =>
-      Localizations.of<Localization>(context, Localization);
+  static Localization of(BuildContext context) => Localizations.of<Localization>(context, Localization)!;
 
-  static Future<Localization> load(Locale locale,
-      {bool showLocalizationKeys = false}) async {
+  static Future<Localization> load(Locale locale, {bool showLocalizationKeys = false}) async {
     final localizations = Localization();
     if (showLocalizationKeys) {
       return localizations;
     }
-    final jsonContent = await rootBundle
-        .loadString('assets/locale/${locale.languageCode}.json');
+    final jsonContent = await rootBundle.loadString('assets/locale/${locale.languageCode}.json');
     // ignore: avoid_as
-    localizations._localisedValues =
-        json.decode(jsonContent) as Map<String, dynamic>;
+    localizations._localisedValues = json.decode(jsonContent) as Map<String, dynamic>;
     return localizations;
   }
 
-  String _t(String key, {List<dynamic> args}) {
+  String _t(String key, {List<dynamic>? args}) {
     try {
       // ignore: avoid_as
       var value = _localisedValues[key] as String;
       if (value == null) return '$key';
       if (args == null || args.isEmpty) return value;
-      args
-          .asMap()
-          .forEach((index, arg) => value = _replaceWith(value, arg, index + 1));
+      args.asMap().forEach((index, arg) => value = _replaceWith(value, arg, index + 1));
       return value;
     } catch (e) {
       return '⚠$key⚠';
@@ -58,12 +52,10 @@ class Localization {
 
   String testArg2(num arg1) => _t(LocalizationKeys.testArg2, args: [arg1]);
 
-  String testArg3(String arg1, num arg2) =>
-      _t(LocalizationKeys.testArg3, args: [arg1, arg2]);
+  String testArg3(String arg1, num arg2) => _t(LocalizationKeys.testArg3, args: [arg1, arg2]);
 
-  String testArg4(String arg1, num arg2) =>
-      _t(LocalizationKeys.testArg4, args: [arg1, arg2]);
+  String testArg4(String arg1, num arg2) => _t(LocalizationKeys.testArg4, args: [arg1, arg2]);
 
-  String getTranslation(String key, {List<dynamic> args}) =>
-      _t(key, args: args ?? List());
+  String getTranslation(String key, {List<dynamic>? args}) => _t(key, args: args ?? []);
+
 }