+++ example/lib/util/locale/localization_keys.dart
@@ -2,6 +2,7 @@
 //THIS FILE IS AUTO GENERATED. DO NOT EDIT//
 //============================================================//
 class LocalizationKeys {
+
   static const test = 'test';
 
   static const testArg1 = 'test_arg1';
@@ -11,4 +12,5 @@ class LocalizationKeys {
   static const testArg3 = 'test_arg3';
 
   static const testArg4 = 'test_arg4';
+
 }